+++ package.json
@@ -46,6 +46,7 @@
   "dependencies": {
     "ascii-art": "2.8.5",
     "chalk": "4.1.0",
-    "figlet": "1.5.2"
+    "figlet": "1.5.2",
+    "link": "^1.5.1"
   }
 }
