+++ src/__mocks__/welcome.js
@@ -1,0 +1,4 @@
+// eslint-disable-next-line no-undef
+const welcome = jest.fn(() => 'Welcome to md-links!');
+
+module.exports = welcome;
