+++ src/cli.js
@@ -13,11 +13,7 @@ const options = process.argv.slice(2);
 if (options.length === 1) {
   mdLinks(path, { validate: false })
     .then((data) => {
-      data.forEach((link) => {
-        log(message(`Archivo: ${link}`, 'orange'));
-        log(message('Links:', 'green'));
-        log(data);
-      });
+      log(data);
     })
     .catch((error) => {
       log(message(error, 'red'));
