--- test/md-links.spec.js
@@ -1,7 +1,0 @@
-const mdLinks = require('../src/md_links');
-
-describe('mdLinks', () => {
-  it('is a function', () => {
-    expect(typeof mdLinks).toBe('function');
-  });
-});
